Moved resources import function to their own import_rancher2_<resource>.go
