+++ rancher2/resource_rancher2_namespace.go
@@ -383,30 +383,6 @@ func resourceRancher2NamespaceDelete(d *schema.ResourceData, meta interface{}) e
 	return nil
 }
 
-func resourceRancher2NamespaceImport(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
-	clusterID, resourceID := splitID(d.Id())
-
-	client, err := meta.(*Config).ClusterClient(clusterID)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-	err = d.Set("project_id", clusterID)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-	ns, err := client.Namespace.ByID(resourceID)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	err = flattenNamespace(d, ns)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	return []*schema.ResourceData{d}, nil
-}
-
 // namespaceStateRefreshFunc returns a resource.StateRefreshFunc, used to watch a Rancher Namespace.
 func namespaceStateRefreshFunc(client *clusterClient.Client, nsID string) resource.StateRefreshFunc {
 	return func() (interface{}, string, error) {
