+++ rancher2/resource_rancher2_node_driver.go
@@ -335,24 +335,6 @@ func resourceRancher2NodeDriverDelete(d *schema.ResourceData, meta interface{}) 
 	return nil
 }
 
-func resourceRancher2NodeDriverImport(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
-	client, err := meta.(*Config).ManagementClient()
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-	nodeDriver, err := client.NodeDriver.ByID(d.Id())
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	err = flattenNodeDriver(d, nodeDriver)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	return []*schema.ResourceData{d}, nil
-}
-
 // nodeDriverStateRefreshFunc returns a resource.StateRefreshFunc, used to watch a Rancher NodeDriver.
 func nodeDriverStateRefreshFunc(client *managementClient.Client, nodeDriverID string) resource.StateRefreshFunc {
 	return func() (interface{}, string, error) {
