+++ rancher2/resource_rancher2_project_logging.go
@@ -512,24 +512,6 @@ func resourceRancher2ProjectLoggingDelete(d *schema.ResourceData, meta interface
 	return nil
 }
 
-func resourceRancher2ProjectLoggingImport(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
-	client, err := meta.(*Config).ManagementClient()
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-	projectLogging, err := client.ProjectLogging.ByID(d.Id())
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	err = flattenProjectLogging(d, projectLogging)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	return []*schema.ResourceData{d}, nil
-}
-
 // projectLoggingStateRefreshFunc returns a resource.StateRefreshFunc, used to watch a Rancher Cluster Role Template Binding.
 func projectLoggingStateRefreshFunc(client *managementClient.Client, projectLoggingID string) resource.StateRefreshFunc {
 	return func() (interface{}, string, error) {
