+++ rancher2/resource_rancher2_setting.go
@@ -283,24 +283,6 @@ func resourceRancher2SettingDelete(d *schema.ResourceData, meta interface{}) err
 	return nil
 }
 
-func resourceRancher2SettingImport(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
-	client, err := meta.(*Config).ManagementClient()
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-	setting, err := client.Setting.ByID(d.Id())
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	err = flattenSetting(d, setting)
-	if err != nil {
-		return []*schema.ResourceData{}, err
-	}
-
-	return []*schema.ResourceData{d}, nil
-}
-
 // settingStateRefreshFunc returns a resource.StateRefreshFunc, used to watch a Rancher Project.
 func settingStateRefreshFunc(client *managementClient.Client, settingID string) resource.StateRefreshFunc {
 	return func() (interface{}, string, error) {
